refactor - muevo funciones de db a utils.py
